Refactors logging system to use singleton pattern
移除旧日志头文件并引入新的日志头文件，采用单例模式创建Logger类，确保全局唯一实例。更新相关代码以使用新实现，简化日志记录过程，改善代码组织和可读性。
