+++ Source/Core/main.cpp
@@ -4,7 +4,7 @@
 #include <cstdio>
 
 #include "FreeRTOS.h"
-#include "bsp_log.h"
+#include "bsp_log.hpp"
 #include "bsp_uid.h"
 #include "chronolink.h"
 #include "harness.h"
