+++ Source/Core/slave/harness.cpp
@@ -3,9 +3,7 @@
 #include <cstdint>
 #include <cstdio>
 
-#include "bsp_log.h"
-
-extern Logger Log;
+#include "bsp_log.hpp"
 
 Harness harness;
